Remove unnecessary client dependencies
