+++ codalab/objects/permission.py
@@ -1,8 +1,6 @@
 '''
 Defines ORM classes for groups and permissions.
 '''
-from marshmallow import fields, ValidationError
-
 from codalab.model.orm_object import ORMObject
 from codalab.common import (
     NotFoundError,
@@ -23,47 +21,6 @@ from codalab.model.tables import (
 )
 from codalab.model.util import LikeQuery
 
-
-class PermissionSpec(fields.Field):
-    def _serialize(self, value, attr, obj):
-        try:
-            return permission_str(value)
-        except UsageError as e:
-            raise ValidationError(e.message)
-
-    def _deserialize(self, value, attr, data):
-        try:
-            return parse_permission(value)
-        except UsageError as e:
-            raise ValidationError(e.message)
-
-
-# TODO(sckoo): delete when REST API is complete
-class Group(ORMObject):
-    '''
-    Defines a group object which is used to assign permissions to a set of users.
-    '''
-    COLUMNS = ('uuid', 'name', 'owner_id', 'user_defined')
-
-    def validate(self):
-        '''
-        Check a number of basic conditions that would indicate serious errors if
-        they do not hold. Right now, validation only checks this worksheet's uuid
-        and its name.
-        '''
-        spec_util.check_uuid(self.uuid)
-        spec_util.check_name(self.name)
-        precondition(isinstance(self.owner_id, basestring), 'Invalid value: owner_id.')
-        precondition(isinstance(self.user_defined, bool), 'Invalid value: user_defined.')
-
-    def __repr__(self):
-        return 'Group(uuid=%r, name=%r)' % (self.uuid, self.name)
-
-    def update_in_memory(self, row, strict=False):
-        if strict:
-            if 'uuid' not in row:
-                row['uuid'] = spec_util.generate_uuid()
-        super(Group, self).update_in_memory(row)
 
 ############################################################
