+++ requirements.txt
@@ -1,10 +1,6 @@
 argcomplete==1.1.0
-bottle==0.12.9
-PyYAML==3.11
 SQLAlchemy==1.0.8
-alembic==0.8.2
-marshmallow-jsonapi==0.6.0
-oauthlib==1.0.3
+PyYAML==3.11
 psutil==3.3.0
 six==1.10.0
 watchdog==0.8.3
